requirements.txt
Django==5.1.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.4.0
django-filter==24.3
WeasyPrint==62.3
